pyproject.toml
[tool.poetry]
name = "suggestions-collector"
version = "0.1.0"
description = ""
authors = ["Ilia Kaziamov <kaziamov@outlook.com>"]
readme = "README.md"
packages = [{include = "suggestions_collector"}]

[tool.poetry.dependencies]
python = "^3.8.1"
beautifulsoup4 = "^4.11.1"
youtube-search-python = "^1.6.6"
asyncio = "^3.4.3"
aiohttp = "^3.8.3"
google-api-python-client = "^2.82.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
